Fix three Stream-blocker bugs: int64 literals, function ADT fields, 0-arity CAFs
Three bugs surfaced while building Std\Stream, all now fixed.

== Bug 1: Integer literals overflow at INT32_MAX ==

`1000000000000` parsed as `-727379968`. The lexer always read via stoll
(64-bit) but the AST node IntegerExpr was templated on `LiteralExpr<int>`
and the parser cast the 64-bit value to `int` before storing it. Fixed
by making IntegerExpr use int64_t end-to-end and removing the casts in
ParserExpr.cpp / ParserPattern.cpp. Test:
test/codegen/int_literal_int64.yona — 1e12 + 2e12 = 3e12.

== Bug 2: ADT field types beyond head identifier ==

A field declared as `(() -> Stream a)` lost the return type by the time
it reached the call site of the extracted function, so calling
`rest ()` produced a value typed i64 instead of the expected ADT, and
pattern matching the result crashed in CreateExtractValue.

Fixed in three pieces:

  * register_yona_module_decls and compile_module both run a fixpoint
    pass over the module's ADTs at registration time, marking any ADT
    that references another ADT-with-function-fields as recursive (heap
    allocated). This makes mutual references like
      Stream a = Stream (() -> Step a)
      Step a   = Yield a (Stream a) | Done
    correctly use the heap layout for both, instead of letting one of
    them remain a flat-struct ADT that breaks the closure ABI.

  * AdtInfo gains field_fn_return_types and field_fn_return_adt_names,
    populated whenever a field is declared as a function type with a
    known return-type head identifier.

  * codegen_pattern_constructor propagates that return CType / ADT
    name into the bound TypedValue.subtypes for function-typed fields,
    so the higher-order call site picks the right LLVM signature.

  * compile_function's recreate path was leaving a stale Function* in
    compiled_functions_[name] when it had to recreate a function with
    a different return type. Self-recursive calls inside the re-codegen'd
    body then jumped to the freed function. Fixed by updating the
    cached entry before re-codegen.

== Bug 3: 0-arity exported value definitions ==

`naturals = range 0 N` at module top-level resolved as a deferred
function reference (CType::FUNCTION) instead of the value it's meant
to compute, so `take 5 naturals` crashed and `let n = naturals` left
n unbound.

Fixed in codegen_identifier: when an identifier resolves to a deferred
function with zero parameters, compile it and call it with no args,
returning the result. This is the standard Haskell-CAF semantics.
Test: test/codegen/stdlib_naturals_caf.yona.

== Stream module ==

lib/Std/Stream.yona's naturals is now a 0-arity CAF that returns
`range 0 INT64_MAX` directly — no thunk wrapper. Tests added:

  * test/codegen/stream_basic.yona — sum (range 1 11)
  * test/codegen/stream_filter_take.yona — toSeq (take 5 (filter ...))
  * test/codegen/stream_iterate.yona — toSeq (take 6 (iterate (* 2) 1))
  * test/codegen/stream_pipeline.yona — naturals |> map |> filter |> sum
  * test/codegen/stdlib_naturals_caf.yona — sum (take 10 naturals)

== Bugs noted but not fixed ==

Three new bugs surfaced during the Stream work and went into TODO with
repros: pipe operator only accepts named functions (not partial
applications), zip with tuple-typed elements segfaults at runtime
(tuple-in-ADT-field heap tracking), and cons of a tuple element into
a Seq segfaults (same family). All three are blocked on extending the
heap-tracking story to TUPLE/SEQ values stored in generic fields —
the same fix family as bug 2 but applied to non-ADT heap types.

Two new follow-up modules added to TODO: Std\IO (stdin/stdout/stderr,
formatted output, line-oriented read) and Std\Constants (intMax,
intMin, math constants, platform constants — single source of truth
for things like INT64_MAX that currently get hard-coded into stdlib
modules).

+++ include/ast.h
@@ -523,12 +523,12 @@ public:
   [[nodiscard]] AstNodeType get_type() const override { return AST_FLOAT_EXPR; };
 };
 
-class YONA_API IntegerExpr final : public LiteralExpr<int> {
+class YONA_API IntegerExpr final : public LiteralExpr<int64_t> {
 private:
   void print(std::ostream &os) const override;
 
 public:
-  explicit IntegerExpr(SourceContext token, int value);
+  explicit IntegerExpr(SourceContext token, int64_t value);
   template<typename ResultType>
   ResultType accept(const AstVisitor<ResultType> &visitor) const {
     return visitor.visit(const_cast<typename std::remove_const<typename std::remove_pointer<decltype(this)>::type>::type*>(this));
